+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -67,6 +67,7 @@ public let masterRuleList = RuleList(rules:
     OpeningBraceRule.self,
     OperatorFunctionWhitespaceRule.self,
     PrivateOutletRule.self,
+    PrivateUnitTestRule.self,
     ReturnArrowWhitespaceRule.self,
     StatementPositionRule.self,
     TodoRule.self,