+++ CHANGELOG.md
@@ -44,6 +44,10 @@
   [bootstraponline](https://github.com/bootstraponline)
   [#689](https://github.com/realm/SwiftLint/issues/689)
 
+* Add rule to check for private unit tests (private unit tests don't get run 
+  by XCTest).  
+  [Cristian Filipov](https://github.com/cfilipov)
+
 * Add configuration for setting a warning threshold.  
   [woodhamgh](https://github.com/woodhamgh)
   [696](https://github.com/realm/SwiftLint/issues/696)